+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## 12.2.0 (upcoming)
 
+- Add support for `\setplength` (L<sup>A</sup>T<sub>E</sub>X)
 - Bundle Java for platform-dependent, standalone archives (see [#70](https://github.com/valentjn/ltex-ls/issues/70))
 
 ## 12.1.0 (May 1, 2021)